+++ src/main/java/site/hirecruit/hr/domain/user/repository/TempUserRepository.kt
@@ -1,8 +1,8 @@
-package site.hirecruit.hr.domain.auth.repository
+package site.hirecruit.hr.domain.user.repository
 
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
-import site.hirecruit.hr.domain.auth.entity.TempUserEntity
+import site.hirecruit.hr.domain.user.entity.TempUserEntity
 
 @Repository
 interface TempUserRepository : CrudRepository<TempUserEntity, Long>
