+++ src/main/java/site/hirecruit/hr/domain/user/service/TempUserRegistrationService.kt
@@ -2,8 +2,8 @@ package site.hirecruit.hr.domain.user.service
 
 import site.hirecruit.hr.domain.auth.dto.AuthUserInfo
 import site.hirecruit.hr.domain.auth.dto.mapper.AuthUserInfoMapper
-import site.hirecruit.hr.domain.auth.entity.TempUserEntity
-import site.hirecruit.hr.domain.auth.repository.TempUserRepository
+import site.hirecruit.hr.domain.user.entity.TempUserEntity
+import site.hirecruit.hr.domain.user.repository.TempUserRepository
 import site.hirecruit.hr.domain.user.dto.TempUserRegistrationDto
 
 /**
