+++ src/test/java/site/hirecruit/hr/domain/auth/service/impl/UserSessionAuthServiceImplTest.kt
@@ -13,8 +13,8 @@ import org.springframework.security.oauth2.core.OAuth2AuthenticationException
 import site.hirecruit.hr.domain.auth.dto.AuthUserInfo
 import site.hirecruit.hr.domain.auth.dto.OAuthAttributes
 import site.hirecruit.hr.domain.user.entity.Role
-import site.hirecruit.hr.domain.auth.entity.TempUserEntity
-import site.hirecruit.hr.domain.auth.repository.TempUserRepository
+import site.hirecruit.hr.domain.user.entity.TempUserEntity
+import site.hirecruit.hr.domain.user.repository.TempUserRepository
 import site.hirecruit.hr.domain.user.repository.UserRepository
 import kotlin.random.Random
